warning fix after huge file patch
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@29443 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/intl.cpp
@@ -1120,7 +1120,7 @@ bool wxMsgCatalogFile::Load(const wxChar *szDirPrefix, const wxChar *szName0,
     return false;
 
   // get the file size
-  off_t nSize = fileMsg.Length();
+  size_t nSize = fileMsg.Length();
   if ( nSize == wxInvalidOffset )
     return false;
